+++ R/usmap.R
@@ -3,7 +3,7 @@
 #' @description
 #' It is usually difficult or inconvenient to create US maps that
 #'   include both Alaska and Hawaii in a convenient spot. All map
-#'   data frames produced by this package use the Alber's Equal Area
+#'   data frames produced by this package use the Albers Equal Area
 #'   projection.
 #'
 #' @section Map data frames:
@@ -69,11 +69,17 @@
 #' @param regions The region breakdown for the map, can be one of
 #'   (\code{"states"}, \code{"state"}, \code{"counties"}, \code{"county"}).
 #'   The default is \code{"states"}.
-#' @param include The regions to include in the output data frame. If \code{regions} is
+#' @param include The regions to include in the resulting map. If \code{regions} is
 #'  \code{"states"}/\code{"state"}, the value can be either a state name, abbreviation or FIPS code.
 #'  For counties, the FIPS must be provided as there can be multiple counties with the
 #'  same name. If states are provided in the county map, only counties in the included states
 #'  will be returned.
+#' @param exclude The regions to exclude in the resulting map. If \code{regions} is
+#'  \code{"states"}/\code{"state"}, the value can be either a state name, abbreviation or FIPS code.
+#'  For counties, the FIPS must be provided as there can be multiple counties with the
+#'  same name. The regions listed in the \code{include} parameter are applied first and the
+#'  \code{exclude} regions are then removed from the resulting map. Any excluded regions
+#'  not present in the included regions will be ignored.
 #'
 #' @return A data frame of US map coordinates divided by the desired \code{regions}.
 #'
@@ -82,9 +88,12 @@
 #'
 #' df <- us_map(regions = "counties")
 #' west_coast <- us_map(include = c("CA", "OR", "WA"))
+#'
+#' south_atl_excl_FL <- us_map(include = .south_atlantic, exclude = "FL")
 #' @export
 us_map <- function(regions = c("states", "state", "counties", "county"),
-                   include = c()) {
+                   include = c(),
+                   exclude = c()) {
 
   regions_ <- match.arg(regions)
 
@@ -96,7 +105,17 @@ us_map <- function(regions = c("states", "state", "counties", "county"),
                         stringsAsFactors = FALSE)
 
   if (length(include) > 0) {
-    df <- df[df$full %in% include | df$abbr %in% include | df$fips %in% include, ]
+    df <- df[df$full %in% include |
+               df$abbr %in% include |
+               df$fips %in% include |
+               substr(df$fips, 1, 2) %in% include, ]
+  }
+
+  if (length(exclude) > 0) {
+    df <- df[!(df$full %in% exclude |
+                 df$abbr %in% exclude |
+                 df$fips %in% exclude |
+                 substr(df$fips, 1, 2) %in% exclude), ]
   }
 
   df